@@ -13,25 +13,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.1",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "13.3.1",
-    "@angular/compiler": "13.3.1",
-    "@angular/core": "13.3.1",
-    "@angular/forms": "13.3.1",
-    "@angular/platform-browser": "13.3.1",
-    "@angular/platform-browser-dynamic": "13.3.1",
-    "@angular/platform-server": "13.3.1",
-    "@angular/router": "13.3.1",
-    "@fortawesome/angular-fontawesome": "^0.9.0",
-    "@fortawesome/fontawesome-svg-core": "1.2.30",
-    "@fortawesome/free-solid-svg-icons": "5.14.0",
-    "@ng-select/ng-select": "^8.3.0",
-    "@ngrx/effects": "^13.2.0",
-    "@ngrx/schematics": "^13.2.0",
-    "@ngrx/store": "^13.2.0",
-    "@ngrx/store-devtools": "^13.2.0",
-    "@types/moment-timezone": "^0.5.30",
+    "@angular/animations": "14.2.1",
+    "@angular/common": "14.2.1",
+    "@angular/compiler": "14.2.1",
+    "@angular/core": "14.2.1",
+    "@angular/forms": "14.2.1",
+    "@angular/platform-browser": "14.2.1",
+    "@angular/platform-browser-dynamic": "14.2.1",
+    "@angular/platform-server": "14.2.1",
+    "@angular/router": "14.2.1",
+    "@fortawesome/angular-fontawesome": "^0.11.1",
+    "@fortawesome/fontawesome-svg-core": "^6.2.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.0",
+    "@ng-select/ng-select": "^9.0.2",
+    "@ngrx/effects": "^14.3.1",
+    "@ngrx/schematics": "^14.3.1",
+    "@ngrx/store": "^14.3.1",
+    "@ngrx/store-devtools": "^14.3.1",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.3.1",
     "chart.js": "^3.8.0",
@@ -40,11 +38,11 @@
     "guid-typescript": "^1.0.9",
     "jquery": "3.5.0",
     "jsencrypt": "^3.2.1",
-    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
     "moment": "^2.29.3",
     "moment-timezone": "^0.5.34",
     "ng-recaptcha": "^9.0.0",
-    "ng-zorro-antd": "^13.4.0",
+    "ng-zorro-antd": "^14.0.0",
     "ng2-charts": "^3.0.11",
     "ngx-bootstrap": "^5.2.0",
     "oidc-client": "^1.9.0",
@@ -53,10 +51,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.1",
-    "@angular/cli": "^13.3.7",
-    "@angular/compiler-cli": "13.3.1",
-    "@angular/language-service": "13.3.1",
+    "@angular-devkit/build-angular": "^14.2.2",
+    "@angular/cli": "^14.2.2",
+    "@angular/compiler-cli": "14.2.1",
+    "@angular/language-service": "14.2.1",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "^12.11.1",
