@@ -7,7 +7,7 @@ import {
   OnInit,
   Output,
 } from "@angular/core";
-import { FormBuilder, FormGroup, Validators } from "@angular/forms";
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from "@angular/forms";
 import { Store } from "@ngrx/store";
 import { NzModalRef } from "ng-zorro-antd/modal";
 import { saveWeatherData } from "src/app/store/weather/weather.actions";
@@ -31,14 +31,14 @@ export class EditWeatherDataComponent implements OnInit, AfterViewChecked {
   @Input() weatherData: DailyForecastWeatherDto;
   @Input() country: CountryDto;
   @Input() province: ProvinceDto;
-  editWeatherForm: FormGroup;
+  editWeatherForm: UntypedFormGroup;
   debug = false;
   weatherConditions$ = this.store.select(selectWeatherConditions); //WeatherConditionCollectionDto;
   loadingSaveWeatherData$ = this.store.select(selectLoadingSaveWeatherData);
 
   constructor(
     private modal: NzModalRef,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private store: Store,
     private readonly changeDetectorRef: ChangeDetectorRef
   ) {}
