@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from "@angular/core";
-import { FormBuilder, FormGroup, Validators } from "@angular/forms";
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from "@angular/forms";
 import { Store } from "@ngrx/store";
 import { NzModalRef } from "ng-zorro-antd/modal";
 import { NoWhitespaceValidator } from "src/app/validators/no-whitespace.validator";
@@ -15,13 +15,13 @@ import { saveCountry } from "src/app/store/location/location.actions";
 })
 export class AddEditCountryModalComponent implements OnInit {
   @Input() country: CountryDto;
-  editForm: FormGroup;
+  editForm: UntypedFormGroup;
   debug = false;
   timezones = [];
   loadingSaveCountry$ = this.store.select(selectLoadingSaveCountry);
   constructor(
     private modal: NzModalRef,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private store: Store
   ) {}
 
